Clean up and break into new component file
